+++ code/qcommon/vm_x86.c
@@ -217,6 +217,7 @@ static qint jumpSizeChanged;
 #endif
 
 static qint funcOffset[FUNC_LAST];
+static qbool forceDataMask;
 
 //literal pool
 #if defined(USE_LITERAL_POOL)
@@ -2406,7 +2407,7 @@ EmitCallOffset(func_t Func)
 static void
 emit_CheckReg(vm_t *vm, uint32_t reg, func_t func)
 {
-  if (vm->forceDataMask || !(vm_rtChecks->integer & VM_RTCHECK_DATA))
+  if (forceDataMask)
   {
 #if idx64
     emit_and_rx(reg, R_DATAMASK); //reg = reg & dataMask
@@ -2898,7 +2899,8 @@ ConstOptimize(vm_t *vm, instruction_t *ci, instruction_t *ni)
       }
       else
       {
-        qint rx = load_rx_opstack(R_EAX | RCONST); dec_opstack(); //eax = *opstack; opstack -= 4
+        qint rx = load_rx_opstack(forceDataMask ? R_EAX:R_EAX | RCONST);
+        dec_opstack(); //eax = *opStack; opStack -= 4
         emit_CheckReg(vm, rx, FUNC_DATW);
         emit_store_imm32_index(ci->value, R_DATABASE, rx); //(dword*)dataBase[eax] = 0x12345678
         unmask_rx(rx);
@@ -2929,11 +2931,8 @@ ConstOptimize(vm_t *vm, instruction_t *ci, instruction_t *ni)
       }
       else
       {
-        //eax = *opstack; opstack -= 4
-        qint rx = load_rx_opstack(R_EAX | RCONST);
-
-        dec_opstack();
-
+        qint rx = load_rx_opstack(forceDataMask ? R_EAX:R_EAX | RCONST);
+        dec_opstack(); //eax = *opStack; opStack -= 4
         emit_CheckReg(vm, rx, FUNC_DATW);
         emit_store2_imm16_index(ci->value, R_DATABASE, rx); //(word *)dataBase[eax] = 0x12345678
         unmask_rx(rx);
@@ -2964,11 +2963,8 @@ ConstOptimize(vm_t *vm, instruction_t *ci, instruction_t *ni)
       }
       else
       {
-        //eax = *opstack; opstack -= 4
-        qint rx = load_rx_opstack(R_EAX | RCONST);
-
-        dec_opstack();
-
+        qint rx = load_rx_opstack(forceDataMask ? R_EAX:R_EAX | RCONST);
+        dec_opstack(); //eax = *opStack; opStack -= 4
         emit_CheckReg(vm, rx, FUNC_DATW);
         emit_store1_imm8_index(ci->value, R_DATABASE, rx); //(qchar *)dataBase[eax] = 0x12345678
         unmask_rx(rx);
@@ -3541,6 +3537,15 @@ VM_Compile(vm_t *vm, vmHeader_t *header)
   code = NULL; //we will allocate memory later, after last defined pass
   instructionPointers = NULL;
 
+  if (vm->forceDataMask || (vm_rtChecks->integer & VM_RTCHECK_DATA) == 0)
+  {
+    forceDataMask = qtrue;
+  }
+  else
+  {
+    forceDataMask = qfalse;
+  }
+
   Com_Memset(funcOffset, 0, sizeof(funcOffset));
 
   instructionCount = header->instructionCount;
@@ -3987,7 +3992,7 @@ __compile:
             else
             {
               //address stored in register
-              rx[0] = load_rx_opstack(R_EAX | RCONST); //eax = *opstack
+              rx[0] = load_rx_opstack(forceDataMask ? R_EAX:R_EAX | RCONST); //eax = *opstack
               emit_CheckReg(vm, rx[0], FUNC_DATR);
               sx[0] = alloc_sx(R_XMM0);
               emit_load_sx_index(sx[0], R_DATABASE, rx[0]); //xmmm0 = dataBase[eax]
@@ -4128,8 +4133,14 @@ __compile:
           else
           {
             //address stored in register
-            //rx[0] = rx[1] = load_rx_opstack(R_EAX); //target, address = *opstack
-            load_rx_opstack2(&rx[0], R_EDX, &rx[1], R_EAX); //target, address = *opstack
+            if (forceDataMask)
+            {
+              rx[0] = rx[1] = load_rx_opstack(R_EAX); //target = address = *opStack
+            }
+            else
+            {
+              load_rx_opstack2(&rx[0], R_EDX, &rx[1], R_EAX); //target, address (const) = *opStack
+            }
 
             emit_CheckReg(vm, rx[1], FUNC_DATR); //check address bounds
 
@@ -4212,10 +4223,8 @@ __compile:
             }
             else
             {
-              //edx = *opstack; opstack -= 4
-              rx[1] = load_rx_opstack(R_EDX | RCONST);
-              dec_opstack();
-
+              rx[1] = load_rx_opstack(forceDataMask ? R_EDX:R_EDX | RCONST);
+              dec_opstack(); //edx = *opStack; opStack -= 4
               emit_CheckReg(vm, rx[1], FUNC_DATW);
               emit_store_sx_index(sx[0], R_DATABASE, rx[1]); //dataBase[edx] = xmm0
               unmask_rx(rx[1]);
@@ -4263,10 +4272,8 @@ __compile:
             else
             {
               //address specified by register
-              //edx = *opstack; opstack -= 4
-              rx[1] = load_rx_opstack(R_EDX | RCONST);
-              dec_opstack();
-
+              rx[1] = load_rx_opstack(forceDataMask ? R_EDX:R_EDX | RCONST);
+              dec_opstack(); //edx = *opStack; opStack -= 4
               emit_CheckReg(vm, rx[1], FUNC_DATW);
 
               switch(ci->op)